Move routing to App component only
