+++ src/containers/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-  Route, withRouter,
+  BrowserRouter as Router, Route, Switch , Redirect,
 } from 'react-router-dom';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
@@ -9,12 +9,15 @@ import Header from '../components/Header';
 import FooterNav from '../components/FooterNav';
 import AddMeasure from './AddMeasure';
 import {
-  loadCategoriesList, loadProgress, logOutUser,
+  loadCategoriesList, loadProgress, logOutUser, loginUser
 } from '../actions';
 import expensifyApi from '../api/expensify';
 import Progress from './Progress';
 import ProgressDetail from './ProgressDetail';
 import Authentication from '../utils/authentication';
+import Login from '../components/Login';
+import SignUp from '../components/SignUp';
+
 
 class App extends React.Component {
   constructor(props) {
@@ -48,34 +51,52 @@ class App extends React.Component {
 
   render() {
     const {
-      headerTitle, headerType, activeTab, user,
+      headerTitle, headerType, activeTab, user, loginUser,
       prevPage, updateCategoriesList,
     } = this.props;
 
+    const loggedIn = Authentication.isValidLogin(user, loginUser);
+
     return (
-      <div>
-        <Header
-          headerTitle={headerTitle}
-          headerType={headerType}
-          logOutUser={this.logoutUserApp}
-          prevPage={prevPage}
-        />
-        <div className="contentWrapper">
-          <Route exact path="/expense/:id">
-            <AddMeasure updateTitle={this.updateTitle} loadProgress={this.loadUserProgress} />
+      loggedIn === 'trying' ? <div></div> : ( 
+      <Router>
+        <Switch>
+          <Route exact path="/login">
+            { !loggedIn ? <Login loginUser={loginUser} /> : <Redirect to="/" />}
           </Route>
-          <Route exact path="/progress/month/:month">
-            <ProgressDetail userId={user.id} updateCategoriesList={updateCategoriesList} />
+          <Route exact path="/signUp">
+            { !loggedIn ? <SignUp loginUser={loginUser} /> : <Redirect to="/" />}
           </Route>
-          <Route exact path="/progress/:id">
-            <Progress user={user} loadProgress={this.loadUserProgress} />
+          <Route path="/">
+            { loggedIn ? (
+              <div>
+                <Header
+                  headerTitle={headerTitle}
+                  headerType={headerType}
+                  logOutUser={this.logoutUserApp}
+                  prevPage={prevPage}
+                />
+                <div className="contentWrapper">
+                  <Route exact path="/expense/:id">
+                    <AddMeasure updateTitle={this.updateTitle} loadProgress={this.loadUserProgress} />
+                  </Route>
+                  <Route exact path="/progress/month/:month">
+                    <ProgressDetail userId={user.id} updateCategoriesList={updateCategoriesList} />
+                  </Route>
+                  <Route exact path="/progress/:id">
+                    <Progress user={user} loadProgress={this.loadUserProgress} />
+                  </Route>
+                  <Route exact path="/">
+                    <CategoryList updateTitle={this.updateTitle} />
+                  </Route>
+                </div>
+                <FooterNav actualTab={activeTab} userId={user.id} />
+              </div>
+            ) : <Redirect to="/login" />}
           </Route>
-          <Route exact path="/">
-            <CategoryList updateTitle={this.updateTitle} />
-          </Route>
-        </div>
-        <FooterNav actualTab={activeTab} userId={user.id} />
-      </div>
+        </Switch>
+      </Router>
+      )
     );
   }
 }
@@ -91,6 +112,7 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch => ({
   updateCategoriesList: categories => dispatch(loadCategoriesList(categories)),
   loadProgress: progress => dispatch(loadProgress(progress)),
+  loginUser: data => dispatch(loginUser(data)),
   logOutUser: () => dispatch(logOutUser()),
 });
 
@@ -103,8 +125,9 @@ App.propTypes = {
   headerTitle: PropTypes.string.isRequired,
   headerType: PropTypes.number.isRequired,
   activeTab: PropTypes.string.isRequired,
+  loginUser: PropTypes.func.isRequired,
   logOutUser: PropTypes.func.isRequired,
   prevPage: PropTypes.string.isRequired,
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(withRouter(App));
+export default connect(mapStateToProps, mapDispatchToProps)(App);
