+++ src/index.js
@@ -4,10 +4,11 @@ import './assets/styles/index.scss';
 import { Provider } from 'react-redux';
 import store from './store';
 import SecurityWrapper from './containers/SecurityWrapper';
+import App from './containers/App';
 
 ReactDOM.render(
   <Provider store={store}>
-    <SecurityWrapper />
+    <App />
   </Provider>,
   document.getElementById('root'),
 );
